Fix GSI bug that can drop final client data during SSL handshake. #7332
When doing the SSL handshake in gss_init_sec_context(), if
globus_i_gsi_gss_handshake() indicates the handshake is complete while
there's a final data payload to send, the data is silently dropped.
This puts the client and server out-of-sync, dooming communication on
the channel.

gss_init_sec_context() is checking output_token->length to see if
there's a payload to send before the data has been extracted into that
buffer.

Here, we patch the Globus external that we build for HTCondor.