Fix bug in CEDAR GSI auth where client missed handshake failure. #7332
After the GSI/SSL handshake, the server sends an integer to the client
indicating whether it succeeded, which the client receives into the
local variable 'status'. On a communication failure, the client assumes
that the value of 'status' will be unchanged. This may not be true if
the end_of_message() is what failed.

We observed this when GSI got out-of-sync and the client was reading GSI
handshake data as the post-handshake response. A non-zero value was
written into 'status', end_of_message() failed due to unread data in the
CEDAR packet, and the client proceeded as if authentication succeeded.
Now, we ensure 'status' is set to 0 on a communication failure.