Patch blahp to skip proxy symlinks on refresh when not needed. #7825
On a BLAH_JOB_REFRESH_PROXY command, the blahp no longer complains about
not finding a symlink to the old proxy under $(HOME)/.blah_jobproxy_dir/
if it doesn't need to find the file. The symlink is needed whe the blahp
is switching uids, creating a limited proxy, or forwarding the proxy to
the worker node.