Make DEACTIVATE_CLAIM command more robust to failure. #7840
The client now requires a response from the startd. With shared port and
resuming a security session, the client can send the request and the
startd may never receive it. The response from the startd is the only
way for the client to know if the request was received and acted upon.
If a request from the shadow is not received, the shadow will wait
forever without retrying.

This breaks compatibility with pre-7.0.5 startds. That's old enough to
not be a concern.