Remove redundant sysctl() calls in procapi for macos. #7851
(This was back-ported from #7708 in 8.9.8)
Save information from initial sysctl() call that returns information
about all processes, to avoid subsequent per-process calls of sysctl().
Similar to Windows, this involves a custom buildProcInfoList() and no
buildPidList() call.

On macos, getProcInfoRaw() and the new getProcInfo_impl() recognize when
the caller has pre-filled the data that comes from sysctl().