Fix proxy lifetime for Condor-C proxy refresh command. #7856
Allow the gridmanager to set the lifetime of the proxy delegated for the
CONDOR_JOB_REFRESH_PROXY c-gahp command. Without this, the gridmanager
and c-gahp can have different ideas of what the lifetime will be
(c-gahp doesn't know about DelegateJobGSICredentialsLifetime in the job
ad). This could lead to the gridmanager issuing future refresh commands
at the wrong time.

The new argument to CONDOR_JOB_REFRESH_PROXY is optional. If it's not
given, then the old behavior is done in the c-gahp.