Add MIN_FLOCK_LEVEL config knob. #7881
This lets the admin set the minimum value for submitters' FlockLevel
(the minimum was always 0 before). Consequently, some pools in the
FLOCK_TO list will always be flocked to.

Also, fix a bug that caused a new submitter with idle jobs to always
start flocking (because the schedd erroneously thinks it hasn't heard
from the local negotiator for a long time).