Docker build containers use NAT network to be more reliable
Occasionally, the myhostname service doesn't seem to work with
IPv6 addresses. So, the build fails to get off the ground, because
HTCondor can't deal with it. So, we use the default NAT networking
for build jobs and use the host networking for test jobs to be
able to run the sinful test with IPv6. An occasional test run
failure is much more tolerable.