Do not try FS auth from different hosts. HTCONDOR-87
This causes the client to insert its hostname into the handshake
ClassAd; when the server reconciles the handshake, it will review
the client's hostname against its current hostname.  If there
is a mismatch, then FS is removed from the list of potential
authentication methods suggested by the server.

To avoid this behavior, set `SEC_FS_SKIP_HOSTNAME_CHECK` to `true`.