Always try other auth methods if security library load fails. HTCONDOR-47
If loading of an auth method's shared library fails, CEDAR would skip
all subsequent auth methods in the security negotaition.
In Authentication::handshake_continue(), always look at additional auth
methods if loading the shared library for our top choice fails.
Before, we only did this for GSI. SSL and Kerberos libraries are part of
the OS install, so those are unlikely to trigger this bug.
But with Munge and SciTokens, a missing library is likely.
This is especially imporant if we want to include these methods in the
default list.