Add WIP code for building a daily image and building the image for the "latest" htcondor
Daily builds are tagged with when the _image_ was made, not the RPM
(this makes the tags more predictable), but the CondorVersion labels
have been split up so it's easier to get the build info later.

A bunch of changes were needed to get derived images to work:

- The Dockerfiles for the derived images now take a BASETAG argument for
  their FROM image, e.g. `--build-arg BASETAG=8.9.9-el7` will result in
  `FROM htcondor/base:8.9.9-el7`

- The build script has been adjusted to work with that and now builds
  _all_ derived images, not just the one(s) given on the command line.
  `--mini/--nomini` control whether to build the `mini` (minicondor)
  image, since that one takes a while.  (Off by default, will change
  that when ready for production).

Also add --push and --nopush instead of having to specify a command
Also make the image labels fit the schema better and unset a few more
that the parent might have.

The non-daily EL Dockerfile now also uses
https://research.cs.wisc.edu/htcondor/repo/${SERIES}/htcondor-release-current.el${EL}.noarch.rpm
for the repo definitions.