Initial code for Lumberjack job queue export feature. HTCONDOR-179
The new method ExportJobs() will write a job_queue.log in the given
location that contains just the ads for the given cluster ids.
Many assumptions about the jobs are made but not enforced yet.
In the exported job ads, the base SPOOL directory is changed to '##' and
LeaveJobInQueue is set to true.
In the original job queue, the jobs are marked as externally managed.
No files beyond job_queue.log are handled.
There's no command handler yet. Export is triggered by config params set
at reconfig time.