AES-GCM: Always enable encryption.
For AES-GCM, encryption is at the per-message level whereas prior
methods allowed messages to be a mix of encrypted and unencrypted.

Because we don't necessarily know whether the message *should* be
encrypted on the recieve side, we don't want to add significant state
tracking, and we want to encourage encryption anyway, we made the
executive decision that AES-GCM sockets should always start off life
as encrypted.