Filter crypto methods in client. HTCONDOR-248
This causes the client to fail the authentication if the server
selected a crypto method that is not supported; further, the client
will only suggest crypto methods it supports.

If the client can't support the server-selected method (or the server
selects no methods!), then this will generate a reasonable error message.