Allow KeyCacheEntry to hold multiple keys. HTCONDOR-233
KeyCacheEntry can now hold multiple KeyInfo objects as part of a single
security session (single policy, expiration, etc). The intention is to
have one key per crypto protocol for the session (CryptoMethods in the
policy ad can now be a list of methods). The KeyCacheEntry also stores a
preferred protocol, used to select which key to use when starting a
connection as a client.

KeyCacheEntry does not enforce one key per protocol or that key
protocols and entries in the CryptoMethods attribute match.