CryptoMethods now reflects methods we have keys for. HTCONDOR-233
In the security session policy ad, the methods listed CryptoMethods now
match the methods/keys in the sessions KeyInfo object. That is, they are
the crypto methods that can actually be used once the session is
established.
In the resume session ad sent by the client, CryptoMethods contains just
the method the client will use for the new connection.
In both cases, if the session has no crypto methods/keys, then
CryptoMethods is not set.