Move code to create directories in master to after daemon core starts up.
Because creating the directories needs a call to get_condor_uid(), we
cannot do this in main() else the condor_master will never switch to
user condor!  Moving this code to pre_command_sock_init instead, which
should be early enough but late enough that the priv system is setup.
Also fixed a typo on the SPOOL permission.  HTCONDOR-73