AES-GCM: Do not reset header when we talk to the shared port #6226
We typically have to reset the message digest for headers when
shared port is in use (since the shared port daemon consumes the
first CEDAR message and then hands the socket to the 'real' daemon).

However, if we are talking to the shared port daemon itself, we cannot
reset the MD or we get a failure to decrypt!