Add replay attack protection in CEDAR. HTCONDOR-287
This adds a round trip for a random nonce on session resumption.
This makes each valid session resumption unique.  With AES-GCM's
checksumming of the handshake, it prevents replay of a prior session.

The replay attack protection is negotiated - providing client/servers
with the ability to avoid it.  Currently, the default is to make it
optional.  However, the implementation is non-blocking -- anything
that is latency-sensitive should be already non-blocking.