nsenter: Use a buffer to forward data through TTY FDs.
Previousy, a single char has been used as buffer.
When producing lots of output, this causes high CPU load
of both the sshd process and nsenter,
and up to two cores can be fully loaded by cat'ting a large file.

This commit introduces a small character buffer on the stack,
which allows to bunch these operations. CPU usage for both nsenter
and sshd drops from 100 % to a few %, below the CPU usage of "cat",
and standard output is not CPU bound anymore.