(HTCONDOR-323) Only invoke update callback once.
In the event of a failure to to update the collector, the
update callback was potentially invoked twice: once on failure
to update and a second time when re-establishing a new TCP socket
to the collector.  This caused the (crazy fragile) Token Request
data to be double-freed, leading to a likely daemon crash.