Keep CEDAR max packet size at least 4096 bytes. HTCONDOR-334
The AES encryption code was accidentally reducing the max packet size to
the size of the last packet sent, but allowing it to increase by 16
bytes on each subsequent larger packet.
Now, the max packet size is always at least 4096 (as for non-AES), but
it can still grow by 16 bytes per packet for large messages. Fixing the
slow growth will wait for later.