make lsb_release grep slightly more specific
I don't like the pattern "id" for the grep.  it can match anywhere in
the line, including part of a word, and can match the value rather than
the key.  Specifying '\<id:' forces it to be the last word in the key,
though I assume it would be clearer just to search for "Distributor ID:"
explicitly.

I am not sure though why you don't just run 'lsb_release' with '--id'
and '--codename' though.