Avoid using `apply` function that does not exist in Python3. HTCONDOR-405
This converts a variadic function call using `apply` to simply invoking
the plugin with a fixed number of python arguments (which it will always
have here).

This also wraps the error handling in a try/catch block, preventing
it from propagating exceptions back to the collector (which will then
crash the collector).