(HTCONDOR-434) Fix claim activation failures.
When a job uses job_machine_attrs (MachineAttrXXX) in the
requirements expression, which is a common 'black hole' policy,
the shadow's request to activate the job is often refused by the
startd.  The problem is the schedd modifies the job ad by rotating all
of the MachineAttrXXX values before launching the shadow, resulting in
the job ad being different than it was when it was used for matchmaking
and claiming the slot — and thus when the shadow sends the (now
modified) ad to the startd to activate the claim, it may no longer
match.

To fix, we no longer insert and rotate the MachineAttrXXX values into
the job ad before launching the shadow.  Instead, we only insert
MachineAttrName0 values (the first index) before launching the shadow,
and then rotate the values after the shadow exits.  In addition, we only
rotate if the job is not in a terminal state (completed or removed),
when we rotate we remove the MachineAttrName0 value (since the job is
not currently running), and only rotate if the first index value was
successfully inserted.