Add libcollector_python3_plugin.so symlink HTCONDOR-399
Also added libclassad_python3_user.so symlink since it's in a similar situation.

Both so-files are built with cpython's sugaested suffix that can be used when
importing modules from a directory containing multiple versions pertaining to
the versions of Python the modules were built it. This is new behavior from
HTCondor 8.8, which we just built so-files named "...python3...so", which has
broken some code, so I've added symlinks to the newly named files from the old
names.