Don't create useless MD5 structs in CEDAR socket setup. HTCONDOR-492
When configuring a ReliSock with the negotiated security settings,
initialize encryption before integrity. When integrity is done first and
AES is used, the ReliSock code sets up a pair of MD5 contexts, then
tears them down when it learns AES is being used.

In addition to being useless work, this break FIPS compliance.