Implement trust-on-first-use mode for SSL. HTCONDOR-501
With this, when BOOTSTRAP_SSL_SERVER_TRUST is set to true (defaults
to false), we save the remote CA certificate in the chain to the
local disk after an authentication failure.

This causes the next SSL connection to possibly succeed.

This works nicely with token auto-request mode as the second SSL
connection is the token request.