Implement a new version of private attributes.
Now, any attribute prefixed with `_condor_priv` will be considered
a private attribute.

This includes logic to automatically exclude sending a new-style
private attribute to a daemon that is too old to understand the
prefix.  This avoids us accidentally leaking private attributes to
older daemons that won't treat them carefully.