Reduce file tested from 200Mb to 2Mb to speed test HTCONDOR-536
This test creates a 200 Mb file and the md5 checksum thereof.
On my desktop, creating this takes less than one second.  In
batlab, with real spinning rust disks, this takes between
4 seconds and 6 minutes.  This test also has a five minute
timeout to do file transfer and run the job, and when it
takes minutes to read or write the file from/to disk, the
test fails.  To fix this problem, we can just use a smaller
file.