Merge DelegationProvider and DelegationConsumer classes. HTCONDOR-519
This gives us a convenient object for holding all of the data that's
tied to a proxy certificate (and comes out of a proxy file). This will
help in replacing the GSI functions beyond those for delegation.
The new class is X509Credential and can be used on both sides of the
delegation operation.