Fix HAD finding local Replication dameon address. HTCONDOR-625
When the HAD daemon alters the Sinful of its address to look like the
local Replication daemon's Sinful, it needs to change the port for all
of the addresses in the Sinful. Otherwise, if the HAD and Replication
daemons' Sinfuls have different primary addresses (though both for the
local machine), it will incorrectly conclude that the Replication daemon
isn't on the local machine.