Add option to require client credentials for SSL auth. HTCONDOR-236
In CEDAR, we can now require the client to present a valid credential as
part of SSL authentication. If the client fails to do so, the SSL
authentication fails and other authentication methods will be attempted.
This is controlled by the new config param
AUTH_SSL_REQUIRE_CLIENT_CERTIFICATE, which defaults to false (which
preserves the old behavior of allowing anonymous clients).
Both client and server check this param and will cause the SSL
authentication to fail if a client credential is not available.
This does not apply to SciTokens authentication, which shares the same
code.