Don't hold job if execute directory disappears. HTCONDOR-664
If the temporary execute directory created by the starter disappears
unexpectedly, then any failure is likely due to that rather than a fault
in the job. Don't tell the shadow to put the job on hold in this
situation.

We saw this happen to some glideins, when the host batch system deleted
the glidein's directory without first killing all of the glidein's
processes.