Modernize blahp file layout on bosco_cluster install. HTCONDOR-713
When assembling the blahp files to rsync onto the cluster submit node,
copy the files from their actual locations in the downloaded tarball
instead of from the maze of symlinks under libexec/glite/ (that maze is
about to disappear).
One side effect of this is that the blahp libexec and bin directories
are no longer merged together.