Simplify locating the blahp and retire GLITE_LOCATION. HTCONDOR-713
Use BLAHPD_LOCATION instead of GLITE_LOCATION. The blahp reacts the same
way to both. Also change the default value to $(RELEASE_DIR) and stop
making a tree of symlinks under $(LIBEXEC)/glite for the tarball.
If the value is "/usr", then don't set the environment variable, as this
indicates a native package installation that doesn't require the
environment variable.