Improve startd signaling of half-reaped starters. HTCONDOR-737
When the startd fails to send a signal to a starter, check if the
starter process has been reaped by DaemonCore but the reaper callback
hasn't fired yet. In that case, pretend the signal was successfully
delivered. The startd already does this for the now-defunct old starter.
If sending of the signal fails, the startd acts as if the starter is
completely gone, which can lead to some undesireable actions.