Fix crash in startd pslot preemption code. HTCONDOR-737
Resuorce::kill_claim() can delete its own Resource object in some
situations, in which case, the pslot preemption code shouldn't reference
the Resource object any further, else it will crash the startd. The code
thought this only happened when the Resource was in Active activity.
But there are other times it can happen. The code now checks whether the
Resource object is still in the list of Resources managed by ResMgr.