Handle missing execute syscall during network disconnect. HTCONDOR-752
If the shadow gets a syscall that indicates the job is executing without
seeing a begin_execution syscall, invoke the begin_execution logic.
This ensures the job ad and job event log are updated properly to
indicate the job began running.
The lack of a begin_execution syscall can happen if the starter tried to
send it while the daemons were disconnected.