HTCONDOR-776: Assume POSIX 2008 and C99.
This removes `ifdef` and cmake tests for functionality that was
standardized at least in POSIX 2008 and C99.  Any Unix platform
not conforming to either of those may be broken.

In some cases, this also removes dead code; in others, the ifdef's
were kept to avoid the code being executed on Windows.