Add support for archive transfer.
With this, any file uploads (e.g., `transfer_input_files`) with the
prefix of `tar://` will be automatically decompressed and unpacked
on the remote side.

Internally, instead of using a file transfer plugin, the uploader
sends the bytes over the socket and the downloader will use libarchive
to unpack and write out files.