Remove the term BOSCO from user-facing tools and text. HTCONDOR-733
We are purging the term BOSCO from the HTCondor software. This first
step removes it from almost all user-facing places. Executables are
either deleted or renamed. Tool output has been rephrased.
A bosco native package still exists for the campus factory files, whose
ultimate fate has yet to be determined.
Internal code and files the user is not expected to interact with still
contain the term 'bosco'. Changing the filenames will require some care
to not break existing installations.
For now, leave bosco_cluster as a symlink to the new name
condor_remote_cluster, to ease the transition for existing users.