Fix schedd's rounding of d-slot Disk value. HTCONDOR-798
Since HTCondor 8.4, the startd allocates a minimum of 0.1% of TotalDisk
to each dynamic slot (previously, it was 1%). Update the schedd's
massaging of the p-slot ad to approximate the new d-slot ad to match.
Otherwise, the schedd can overestimate the Disk resource allocated to
the d-slot, leading to it trying to run jobs that the startd will
reject.