Immediately drop privileges in condor_vpn via re-exec.
Onwership of (and access to) `/proc/self/setgroups` will be based
on the UID the `condor_vpn` process was exec'd as.  If `condor_vpn`
was exec'd as root (as is typically done in a rootly condor), then
the various namespace setup routines will crash and burn.

Accordingly, if we were exec'd as root, immediately re-exec as
PRIV_CONDOR.