Remove trailing semicolon in file transfer remaps. HTCONDOR-886
When later code appends more remap rules, this can result in a
double-semicolon, which causes the remap logic to ignore all of the
following rules.
One observed result is that when a job whose UserLog contains directory
elements is spooled, the UserLog is written in the wrong directory when
the output sandbox is transferred from the schedd.