Fix wrong env var paths inside Singularity. HTCONDOR-885
HTCondor publishes several environment variables with the location
to the starter's execute directory.  If running w/ Singularity,
you can change the location where the execute dir is mounted,
meaning these paths need to change.  This patch scans all injected
env vars for the execute dir, and modifies the value of the vars
passed into the container.  This way the paths are correct for
processes running outside the container (like USER_JOB_WRAPPER) and
also inside the container.