HTCONDOR-899 Make cmd_q_protocols test more robust
This test submits a job, then immediately removes it.  In rare cases
the remove while the slot is being claimed.  In this case, the
unclaiming doesn't work, and the slot is stuck in claimed/idle for 10
minutes.  The test then submits a subsequent job which can't match
and the test times out.

To fix this problem, we'll just add a second slot in the test
condor.