Improve shadow handling failure to kill starter during hold. HTCONDOR-933
If the shadow's request to kill the starter fails when trying to put the
job on hold, treat it like a disconnection. Close the syscall socket and
do the normal reconnect logic. If a reconnect succeeds, retry killing
the starter. If killing the starter eventually succeeds or the job lease
expires, then exit with JOB_SHOULD_HOLD.