Hard-kill all shadows on a schedd graceful shutdown. HTCONDOR-960
Jobs without a lease are a rare breed, but we don't want a long
preemption or failures in sending a DEACTIVATE_CLAIM to extend the
schedd's shutdown to the point of other jobs' leases expiring before
restarting.
Still send a DEACTIVATE_CLAIM for lease-less jobs, since older starters
will hang out forever on a disconnect when there's no lease.