HTCONDOR-899 Additional procd debug line
We see a rare failure in the tests where the starter forks
to run a job, and in the child process, before it execs
the job, it registers itself as a subfamily to the procd.
The procd returns an error, rarely, claiming the pid of
the forked starter (which the starter has logged) either
doesn't exist, or isn't in the process tree of its parent.

This diff tells us which of those particular cases is
actually happening, so we can debug it further