Give family/parent/child sessions ADMINISTRATOR access. HTCONDOR-891
The HAD daemon needs ADMINISTRATOR permission to tell the master to turn
the negotiator on and off, and it uses one of these prearranged sessions
to authenticate with the master.

A family of daemons have to trust each other throughly, so adding
ADMINISTRATOR access for inter-family communications should be fine.