HTCONDOR-981 Implement `htcondor annex create`
Adds `htcondor annex create`, which is largely a wrapper around
the code in annex_create.py with options pulled out into commandline
arguments and/or condor config knobs (not yet defined in the param
table as the names are not set in stone).

Adds script files needed by the HPC annex creation script to
$(LIBEXEC)/annex. These must be stored in a predictable location.
These files were also added to the RPM spec.