HTCONDOR-981 Fix option dest variables, add annex create verb
The name of the variables in options are still based on the names
of the flags (--foo -> options.foo), so I overrode these using the
"dest" option for argparse arguments where annex_create is expecting
variables named different from their flag names.

Also added missing verb for annex create and adjusted the order
of verbs as they show up in the annex noun help.