Make User deletion in job router unconditional. HTCONDOR-1005
Once USER_IS_THE_NEW_OWNER=True and sites want to have non-local
identities for User, we'll have to revisit this code. But it's not as
simple as always preserving User in that case.

If the schedd were fixed to correct an inappropriate User value within
the job submission transaction, then we could leave the existing User
value in place in all situations.