Change protection of new remote admin capability attr. HTCONDOR-638
Remove memory sealing code for private attributes in collector's
in-memory ads. Remove special prefix for new private attributes. Protect
the new RemoteAdminCapability attribute in the collector the same way we
protect the Capability attribute of schedd and submitter ads.
In the collector, only return private attributes to a admin-level user
for locate queries (and to the negotiator as before).