Always use replay protection when possible. HTCONDOR-1054
The client now declares if it's expecting a response from the server
when resuming a session. The server respects this declaration. The
client will always request a server response if it believes the server
is capable.
Replay protection now works for non-negotiated sessions, using the
version string in the serialized policy ad and having the server
remember the version of the last client that used each session
(handles case of startd using its non-negotiated session to connect to
an old schedd).