When jobs exhaust their allocated disk space, put them on hold.
This provides a mechanism where jobs will fairly reliably get a
hold message when they exhaust their resources instead of relying on
users correctly managing an "out of space" error in their code.

This is done by thinly allocating an extra 2GB per job; if the 2GB
is used before the issue is noticed, the ENOSPC might occur.  If the
backing pool is exhausted due to the overcommit, then all jobs will
get put into D state on write until the starter's poll cycle notices
this and puts the offending job on hold.  If, for some reason, this
doesn't solve the issue, jobs are evicted and requeued until the
backing store has free space.