Fix python bindings for replay protection on job submit. HTCONDOR-287
The bindings send a RESCHEDULE via resumed session to the schedd while
it has a qmgmt connection open for the job submission. This doesn't work
if the server has to send a reply for the resumed session.
Now, ConnectionSentry in the bindings defers a requested reschedule
until after the open qmgmt connection is closed.