HTCONDOR-1060: Changed gahp and dagman subsystem from daemon to client
DAGMan and gahp susbsytems believe they were daemons when in fact the aren't.
This could lead to issues in authentication where they would pass the subsystem
check of isDaemon() and try to do daemon things. Changed the table to relfect
that they are clients and not daemons to avoid this issue.