HTCONDOR-1054 Allow DCSignalMsg to work without blocking
Clients can now disable the server response to a resume session request
for an individual command. This is done for DCSignalMsg.

Many users of DCSignalMsg send in blocking mode and assume it won't
block when using a cached session. It can now block if an unresponsive
server is expected to send a response to the resumed session.
This can lead to a (timeout-limited) deadlock if the server is trying to
do a blocking connection to this process.
These messages are limited to the local machine and there's no data sent
beyond the command int. Thus, it's safe to disable the server's response
to the resumed session.