HTCONDOR-1083 Store private attributes in separate ad in the collector
The collector now has a private ad chained to each advertised ad. When
an ad update is received, private attributes are moved into the private
ad. Queries are evaluated against the public ad, and the public ad is
usually the one returned. If a client is authorized to see private
attributes, then the private ad is returned (which will include the
chained public ad).
Remove the code to temporarily remove the Capability attribute from each
ad while evaluating the query constraint.