Utilize user-provided hostname, not canonical name, for alias.  HTCONDOR-692
The sinful alias is used in SSL to compare against the SANs listed in the
server certificate.  We should look for whatever the hostname the user provided,
not the canonical name.

The canonical name is actually the name of the A record that any potential CNAME
pointed at; it seems that it is conflated with the FQDN in this particular code.

Instead of trying to guess the actual FQDN - guessing is often a bad idea when the
result is used in security code - we utilize precisely what the user typed in.  This
matches the behavior of `curl` and Google Chrome when presented with a non-FQDN.

The upshot of this change is CEDAR will handle hostnames the way the system administrator
expects them to behave, matching other ocmmon tools.