HTCONDOR-1096 Don't add -lresolv to the global link line
We only use the resolver routine in one place, daemon_core,
where we clear the DNS cache with res_init(), but that function
lives in libc on Linux.  On Mac, we explicitly add it to the
condor_utils dependencies.

With this change, -lresolv isn't added to all our link lines,
which means we can successfully link our static binaries on platforms like el9
which don't currently supply a static version of this library.