HTCONDOR-1125 Make ServerTime attribute an option of putClassAd()
The schedd needs to add a ServerTime attribute to job ads it sends to
condor_q. Before, this was done by having the schedd do this for every
ad (not just job ads) it sends over the network.
Now, adding ServerTime is requested via a flag to putClassAd(), which
the schedd only does for commands QUERY_JOB_ADS and
QUERY_JOB_ADS_WITH_AUTH.