Cleanup of X509 cert creation.  HTCONDOR-501
Ensure we defer creating the new files until the X.509 objects are ready.
If the writing of the object fails, make sure we delete the file we
created.

Note we still don't 'inspect' created files to see if a reasonable
X.509 certificate is inside -- the thinking is that becomes quite
heuristic-driven and race-prone.