HTCONDOR-1163 rename LICENSE-2.0.tx to LICENSE
The Apache licensing guidelines
https://infra.apache.org/apply-license.html
state that the license should be in a file named LICENSE in
the root directory, not the LICENSE-2.0.txt that we have today.
Hopefully with this change, github and other tools will correctly
automatically pick up that HTCondor is licensed with Apache 2.0
(and has been for years).