HTCONDOR-829: WIP code to handle the new feature series versioning scheme
- build-images now provides `9.x` tags if you ask for `9.x` or `latest`;
- container-install-htcondor.sh translates new feature series numbers
  like `9.5`, `9.6`, etc. into `9.x` for repository paths.

The series tags that are applied to the image are still based on the
actual version of condor inside the image -- that is, if you ask for
`--series 9.5` but the latest in that repo is 9.9.0, you will get 9.9.0
and the tags will say `9.9`.  In order to solve it, we'd have to translate
`--series 9.5` into a command like `yum install 'condor>=9.5,condor<9.6'`
and something similar in Ubuntu.  Is it worth the effort, or is 9.5 not
a "real" series?

The right thing _should_ happen if you ask for a specific version like
`--version 9.5.0` but I can't test because of the condor-externals RPM
cicular dependency.