HTCONDOR-1265: Add htcondor/hpc-annex-pilot image
This is a custom image for running the hpc annex pilot itself in; it is
based on Alma 8 since that's what we build EL8 condor with, and includes
condor and singularity (configured to run non-setuid).

The condor RPM might not be necessary if the pilot is using a tarball instead,
but it's the easiest way to make sure we have all the dependencies to run
condor.