Add `DOCKER_SHM_SIZE` knob.
It is desirable to be able to configure Docker's `--shm-size` depending on a job's requested memory or even allow users to specify the value for it in their submission file.
Thus this adds a knob that is parsed as integral value with the machine's and job's classAd available.
If present, the value passed to `--shm-size`.