HTCONDOR-1341 Fix non-negotiated security sessions when using IPv6
When a sinful string has multiple addresses, the Sock object will
rewrite the sinful string so that the preferred address is the primary.
CEDAR then uses the modified sinful string to look for a security session
in the cache. So when creating a non-negotiated session, we need to do the
same rewriting of the sinful string.

This is primarily an issue when PREFER_IPV4=False, where the IPv4 address
will be primary, but the IPv6 address is preferred.