HTCONDOR-1337 Fix memory leak in condor threads
We used to always new up a ThreadImplementation,
then check to see if we are going to create "condor threads"
and if we aren't (the default), destroy the ThreadImplementation,
leaking a thread-local thread id along the way.

This created tons of noise in address sanitizer output,
and was just redundant work.  Let's now check first to
see if we need to create threads before boldly going
down that road.