HTCONDOR-1372 Fix crash in non-blocking connect with session
In SecManStartCommand, when doing a non-blocking connect, don't hold onto
a pointer to the session's KeyCacheEntry object between callbacks. The
object can be removed from the cache and deleted in the mean time, making
later accesses to it invalid (and crash the process).
Instead, copy everything we need (i.e. the session key) when we pull the
KeyCacheEntry out of the cache at the start of the connection process.