HTCONDOR-1426 Fix segfault in curl plugin
Some versions of libcurl, most notably that installed in EL8
pass a char * pointer to a read-only text segment string
to the callback that handles HTTP (and file:) headers.

Our plugin tried to parse these strings with strtok, which
requires writing to them, which resulted in a segfault.

The quick and dirty fix to the problem is to merely copy
the string to the heap, which is not optimal, but fixes
the problem.