HTCONDOR-1427 lib_shared_port-check-ports test fails on ipv6 systems
It counts the number of open TCP ports the shared port has.  It tries
to detect if ipv6 should be in play, but did this by running
condor_config_val IPV6_ENABLED.  This now returns "auto" by default
which isn't useful for the test to decide if it is on or off

Instead, run condor_config_val IPV6_ADDRESS, to see if condor thinks
we have a useful ipv6 address.